Добавил Базу Данных проверил работу приложение с тестового экрана

+++ app/src/main/java/com/example/weatherappdanial/data/remote/dto/Minutely.kt
@@ -6,8 +6,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Minutely(
-    @SerialName("dt")
-    val dt: Int = 0,
-    @SerialName("precipitation")
-    val precipitation: Int = 0
+    @SerialName("dt") val dt: Int = 0,
+    @SerialName("precipitation") val precipitation: Double = 0.0
 )
